core/requirements.txt
psutil>=5.9
rich>=13.0
openai>=1.0
python-telegram-bot>=20.0
send2trash>=1.8
supermemory>=3.0
